requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
django-fcm==0.1.1
django-sites==0.9
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
fcm-django==0.2.11
idna==2.5
olefile==0.44
Pillow==4.2.0
pyfcm==1.3.1
PyJWT==1.5.2
pytz==2017.2
requests==2.18.3
requests-toolbelt==0.8.0
urllib3==1.22
whitenoise==3.3.0
